+++ web/components/admin/body/index.jsp
@@ -1,0 +1,21 @@
+<%-- 
+    Document   : index.jsp
+    Created on : Feb 12, 2024, 3:46:48 PM
+    Author     : ACER
+--%>
+
+<%@page contentType="text/html" pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+        <title>JSP Page</title>
+    </head>
+    <body>
+        <div id="body">
+            <jsp:include page="/components/admin/body/sider/index.jsp" />
+        </div>
+        <jsp:include page="/components/admin/body/mainbody/index.jsp" />
+
+    </body>
+</html>
