+++ web/components/admin/body/mainbody/AccountBody/index.jsp
@@ -1,0 +1,17 @@
+<%-- 
+    Document   : index
+    Created on : Feb 13, 2024, 3:04:24 PM
+    Author     : ACER
+--%>
+
+<%@page contentType="text/html" pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+        <title>JSP Page</title>
+    </head>
+    <body>
+        <h1>Account</h1>
+    </body>
+</html>
