+++ web/components/admin/body/mainbody/RequestBody/index.jsp
@@ -1,0 +1,17 @@
+<%-- 
+    Document   : index
+    Created on : Feb 13, 2024, 3:04:59 PM
+    Author     : ACER
+--%>
+
+<%@page contentType="text/html" pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+        <title>JSP Page</title>
+    </head>
+    <body>
+        <h1>requestBody</h1>
+    </body>
+</html>
