--- web/components/admin/sider/index.jsp
@@ -1,22 +1,0 @@
-<%-- 
-    Document   : SIDER/index
-    Created on : Feb 7, 2024, 11:29:22 AM
-    Author     : ACER
---%>
-
-<%@page contentType="text/html" pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-        <title>SIDER COMP</title>
-    </head>
-    <body>
-        <div id="sider" class="fixed w-[150px] text-white h-[calc(100vh-56px)] top-[56px] left-0 bg-[#323232]">
-            Day la SIDE BAR
-        </div>
-        <div id="main" class="w-[calc(100vw-150px)]">
-            
-        </div>
-    </body>
-</html>
