+++ web/page/loginPage/login.css
@@ -21,7 +21,7 @@
   z-index: 0;
   left: 0;
   width: 100%;
-  background-image: url("../img/login/bg2.png");
+  background-image: url("../../img/login/bg2.png");
   background-position: center;
   background-size: cover;
   height: 100%;
